requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langgraph
requests
beautifulsoup4
trafilatura
python-dotenv
langchain-qdrant
qdrant-client
fastembed